src/document_classifier/requirements.txt
# Required packages for the document classifier
python-magic>=0.4.27
PyPDF2>=3.0.0
pytest>=7.0.0
langchain>=0.0.200
langchain-community>=0.0.10
python-magic-bin>=0.4.14; sys_platform == 'win32'
